+++ Infrastructure/Automation/Mobile/Devices/IDeviceLauncher.cs
@@ -1,0 +1,8 @@
+namespace Infrastructure.Automation.Mobile.Devices;
+
+public interface IDeviceLauncher : IAsyncDisposable
+{
+    Task EnsureStartedAsync();
+
+    Task ShutdownIfStartedAsync();
+}